Address code review comments on admin main menu providers split
