+++ src/Sylius/Bundle/AdminBundle/Menu/Provider/Main/AdministrationMenuProvider.php
@@ -21,7 +21,7 @@ final readonly class AdministrationMenuProvider implements MenuProviderInterface
     public function __invoke(ItemInterface $menu): void
     {
         $administration = $menu
-            ->addChild('sylius.ui.administration')
+            ->addChild('administration')
             ->setLabel('sylius.ui.administration')
             ->setLabelAttribute('icon', 'tabler:lock')
         ;
