+++ src/Sylius/Bundle/AdminBundle/Resources/config/services/menu.php
@@ -17,6 +17,7 @@ use Sylius\Bundle\AdminBundle\Menu\CompositeMenuBuilder;
 use Sylius\Bundle\AdminBundle\Menu\Provider\Main\AdministrationMenuProvider;
 use Sylius\Bundle\AdminBundle\Menu\Provider\Main\CatalogMenuProvider;
 use Sylius\Bundle\AdminBundle\Menu\Provider\Main\ConfigurationMenuProvider;
+use Sylius\Bundle\AdminBundle\Menu\Provider\Main\CustomersMenuProvider;
 use Sylius\Bundle\AdminBundle\Menu\Provider\Main\DashboardMenuProvider;
 use Sylius\Bundle\AdminBundle\Menu\Provider\Main\EventMenuProvider;
 use Sylius\Bundle\AdminBundle\Menu\Provider\Main\MarketingMenuProvider;
@@ -26,14 +27,11 @@ use Sylius\Bundle\AdminBundle\Menu\Provider\Main\SalesMenuProvider;
 return static function (ContainerConfigurator $container) {
     $services = $container->services();
 
-    $services->set('sylius_admin.menu_builder.main', 'sylius_admin.menu_builder.main.composite');
-
     $services
-        ->set('sylius_admin.menu_builder.main.composite', CompositeMenuBuilder::class)
+        ->set('sylius_admin.menu_builder.main', CompositeMenuBuilder::class)
         ->args([
             service('knp_menu.factory'),
             tagged_iterator('sylius_admin.main_menu_provider'),
-            service('router'),
         ])
         ->tag('knp_menu.menu_builder', ['method' => 'createMenu', 'alias' => 'sylius_admin.main'])
     ;
@@ -51,6 +49,11 @@ return static function (ContainerConfigurator $container) {
     $services
         ->set('sylius_admin.menu_builder.provider.main.sales', SalesMenuProvider::class)
         ->tag('sylius_admin.main_menu_provider', ['priority' => -300])
+    ;
+
+    $services
+        ->set('sylius_admin.menu_builder.provider.main.customers', CustomersMenuProvider::class)
+        ->tag('sylius_admin.main_menu_provider', ['priority' => -350])
     ;
 
     $services
